First draft
Converted to md from https://docs.google.com/spreadsheets/d/1B0f1rE5SAkPfOsE0M22QWTDx-z3_gkLvJn6AyVltco0/edit#gid=1364170751
Needs more info filled in (can do in gsheet and copy-paste here), but I needed to have a stable URL for this in GitHub for the paper. Please fix!
BTW, the dots in the URLs seem to be escaped; that probably needs to be fixed.